Remove GAP as misassembly type and related config vars. Rename ERROR misassembly type and rename to COLLAPSE_OTHER. Increase default misjoin valley threshold to require a 95% drop in mean coverage.
