+++ nucflag/classifier/collapse.py
@@ -1,6 +1,65 @@
+import polars as pl
 from collections import defaultdict
 from intervaltree import Interval, IntervalTree
+
+from .common import consecutive
 from ..misassembly import Misassembly
+
+
+def get_secondary_allele_coords(
+    df_cov: pl.DataFrame,
+    *,
+    second_thr: int,
+    group_distance: int,
+    min_group_size: int,
+    thr_het_ratio: float,
+) -> IntervalTree:
+    df_second_outliers = (
+        df_cov.filter(pl.col("second") > second_thr)
+        .with_columns(het_ratio=pl.col("second") / (pl.col("first") + pl.col("second")))
+        .with_columns(
+            het_cls=pl.when(pl.col("het_ratio") > thr_het_ratio)
+            .then(pl.lit("Error"))
+            .otherwise(pl.lit("Het"))
+        )
+    )
+
+    het_coords = IntervalTree()
+    for grp in consecutive(
+        df_second_outliers.filter(pl.col("het_cls") == "Het").get_column("position"),
+        stepsize=group_distance,
+    ):
+        if len(grp) < min_group_size:
+            continue
+        start, end = grp[0], grp[-1]
+        ht = (
+            df_second_outliers.filter(pl.col("position").is_between(start, end))
+            .get_column("second")
+            .median()
+        )
+        het_coords.add(Interval(start, end, (Misassembly.HET, ht)))
+
+    err_coords: IntervalTree = IntervalTree()
+    for grp in consecutive(
+        df_second_outliers.filter(pl.col("het_cls") == "Error").get_column("position"),
+        stepsize=group_distance,
+    ):
+        if len(grp) < min_group_size:
+            continue
+
+        start, end = grp[0], grp[-1]
+        ht = (
+            df_second_outliers.filter(pl.col("position").is_between(start, end))
+            .get_column("second")
+            .median()
+        )
+        err_interval = Interval(start, end, (Misassembly.COLLAPSE_OTHER, ht))
+        # Remove hets.
+        het_coords.chop(err_interval.begin, err_interval.end)
+
+        err_coords.add(err_interval)
+
+    return err_coords.union(het_coords)
 
 
 def identify_collapses(
