+++ nucflag/config.py
@@ -2,7 +2,7 @@ DEF_CONFIG: dict[str, dict] = {
     "first": dict(
         thr_min_peak_width=20,
         thr_min_valley_width=20,
-        thr_misjoin_valley=0.8,
+        thr_misjoin_valley=0.95,
         thr_collapse_peak=2.25,
         valley_group_distance=5_000,
         peak_group_distance=5_000,
@@ -13,6 +13,5 @@ DEF_CONFIG: dict[str, dict] = {
         thr_min_group_size=5,
         thr_het_ratio=0.2,
     ),
-    "gaps": dict(thr_max_allowed_gap_size=0),
     "general": dict(window_size=5_000_000),
 }
