+++ nucflag/misassembly.py
@@ -2,13 +2,12 @@ from enum import StrEnum
 
 
 class Misassembly(StrEnum):
+    COLLAPSE_OTHER = "COLLAPSE_OTHER"
     COLLAPSE_VAR = "COLLAPSE_VAR"
     COLLAPSE = "COLLAPSE"
     MISJOIN = "MISJOIN"
-    GAP = "GAP"
     FALSE_DUP = "FALSE_DUP"
     HET = "HET"
-    ERROR = "ERROR"
 
     def as_color(self) -> str:
         match self:
@@ -18,13 +17,11 @@ class Misassembly(StrEnum):
                 return "green"
             case self.MISJOIN:
                 return "orange"
-            case self.GAP:
-                return "gray"
             case self.FALSE_DUP:
                 return "purple"
             case self.HET:
                 return "teal"
-            case self.ERROR:
+            case self.COLLAPSE_OTHER:
                 return "red"
             case _:
                 raise ValueError(f"Invalid color {self}")
