+++ test/all_gap/expected.bed
@@ -1,1 +1,1 @@
-haplotype2-0000133:3021508-8691473	89881870	96384968	GAP
+haplotype2-0000133:3021508-8691473	89881870	96384968	MISJOIN
