+++ test/misjoin/expected_perc.bed
@@ -1,5 +1,5 @@
 HG00171_chr16_haplotype1-0000003:1881763-8120526	3977787	3978792	HET
-HG00171_chr16_haplotype1-0000003:1881763-8120526	3979637	3991550	ERROR
+HG00171_chr16_haplotype1-0000003:1881763-8120526	3979637	3991550	COLLAPSE_OTHER
 HG00171_chr16_haplotype1-0000003:1881763-8120526	3991550	3994078	HET
 HG00171_chr16_haplotype1-0000003:1881763-8120526	5830988	5852649	MISJOIN
 HG00171_chr16_haplotype1-0000003:1881763-8120526	7287182	7347278	MISJOIN
