+++ test/standard/expected.bed
@@ -1,12 +1,10 @@
 haplotype2-0000133:3021508-8691473	3643515	3660435	HET
-haplotype2-0000133:3021508-8691473	3681959	3715420	ERROR
+haplotype2-0000133:3021508-8691473	3681959	3715420	COLLAPSE_OTHER
 haplotype2-0000133:3021508-8691473	5088022	5091975	HET
-haplotype2-0000133:3021508-8691473	5091975	5100594	ERROR
+haplotype2-0000133:3021508-8691473	5091975	5100594	COLLAPSE_OTHER
 haplotype2-0000133:3021508-8691473	5100594	5105033	HET
 haplotype2-0000133:3021508-8691473	5640379	5645142	HET
-haplotype2-0000133:3021508-8691473	5965734	5971731	ERROR
-haplotype2-0000133:3021508-8691473	7017596	7029969	MISJOIN
-haplotype2-0000133:3021508-8691473	7029969	7146182	GAP
-haplotype2-0000133:3021508-8691473	7146182	7156727	MISJOIN
+haplotype2-0000133:3021508-8691473	5965734	5971731	COLLAPSE_OTHER
+haplotype2-0000133:3021508-8691473	7017596	7156727	MISJOIN
 haplotype2-0000133:3021508-8691473	7570612	7586703	HET
 haplotype2-0000133:3021508-8691473	8388704	8398887	HET
