--- test/subtract_interval/HG00514_chr3_haplotype1-0000046:89839933-96622693.bed.gz
@@ -1,3 +1,0 @@
-version https://git-lfs.github.com/spec/v1
-oid sha256:52babcdebb073b2637ed6e247321f8331421dec0cc9f06cec0ccf4bdad3a8222
-size 16802660
