--- test/subtract_interval/config.toml
@@ -1,47 +1,0 @@
-[first]
-# Min width of peak to consider.
-thr_min_peak_width = 20
-
-# Min width of valley to consider.
-thr_min_valley_width = 20
-
-# Valleys with depth above this threshold are considered misjoins.
-# ex. 0.9
-# Valleys with a depth of 90% of the mean coverage are misjoins.
-thr_misjoin_valley = 0.8
-
-# Peaks with height above this threshold are considered collapses.
-# ex. 2.25
-# Peaks with a height of 225% of the mean coverage are considered collapses
-thr_collapse_peak = 2.25
-
-# Group consecutive positions allowing a maximum gap of x.
-# Larger value groups more positions.
-valley_group_distance = 5000
-
-peak_group_distance = 5000
-
-[second]
-# Percent threshold of most freq base to allow second most freq base
-# ex. 0.1 and a mean coverage of 30.
-# 30 * 0.1 = 3 so above 3 is allowed.
-thr_min_perc_first = 0.05
-
-# Group consecutive positions allowing a maximum gap of x.
-# Larger value groups more positions.
-group_distance = 10_000
-
-# Min group size.
-thr_min_group_size = 5
-
-# Regions with a het ratio below this are hets. Otherwise, considered an error.
-# Lower values will call more het sites as errors.
-thr_het_ratio = 0.2
-
-[gaps]
-# Allow coverage gaps up to this length.
-thr_max_allowed_gap_size = 0
-
-[general]
-# Window size. Only applicable if no region bed file is provided.
-window_size = 5_000_000
