--- test/subtract_interval/expected.bed
@@ -1,24 +1,0 @@
-HG00514_chr3_haplotype1-0000046:89839933-96622693	91049934	91064560	MISJOIN
-HG00514_chr3_haplotype1-0000046:89839933-96622693	91911528	91939397	MISJOIN
-HG00514_chr3_haplotype1-0000046:89839933-96622693	93329010	93337800	ERROR
-HG00514_chr3_haplotype1-0000046:89839933-96622693	93421114	93446040	MISJOIN
-HG00514_chr3_haplotype1-0000046:89839933-96622693	93446040	93446041	GAP
-HG00514_chr3_haplotype1-0000046:89839933-96622693	93446041	93446736	MISJOIN
-HG00514_chr3_haplotype1-0000046:89839933-96622693	93446736	93624704	GAP
-HG00514_chr3_haplotype1-0000046:89839933-96622693	93624704	93625639	MISJOIN
-HG00514_chr3_haplotype1-0000046:89839933-96622693	93625639	93625640	GAP
-HG00514_chr3_haplotype1-0000046:89839933-96622693	93625640	93626555	MISJOIN
-HG00514_chr3_haplotype1-0000046:89839933-96622693	93626555	93626556	GAP
-HG00514_chr3_haplotype1-0000046:89839933-96622693	93626556	93638444	MISJOIN
-HG00514_chr3_haplotype1-0000046:89839933-96622693	93826904	93837417	HET
-HG00514_chr3_haplotype1-0000046:89839933-96622693	93867189	93874111	ERROR
-HG00514_chr3_haplotype1-0000046:89839933-96622693	94075792	94088587	MISJOIN
-HG00514_chr3_haplotype1-0000046:89839933-96622693	94088587	94088589	GAP
-HG00514_chr3_haplotype1-0000046:89839933-96622693	94088589	94094692	MISJOIN
-HG00514_chr3_haplotype1-0000046:89839933-96622693	94206543	94208288	ERROR
-HG00514_chr3_haplotype1-0000046:89839933-96622693	94262047	94283558	MISJOIN
-HG00514_chr3_haplotype1-0000046:89839933-96622693	94857602	94883190	MISJOIN
-HG00514_chr3_haplotype1-0000046:89839933-96622693	95753732	95788229	MISJOIN
-HG00514_chr3_haplotype1-0000046:89839933-96622693	95871277	95888467	MISJOIN
-HG00514_chr3_haplotype1-0000046:89839933-96622693	96198288	96199749	MISJOIN
-HG00514_chr3_haplotype1-0000046:89839933-96622693	96351554	96407315	MISJOIN
