--- test/subtract_interval/region.bed
@@ -1,1 +1,0 @@
-HG00514_chr3_haplotype1-0000046	89839933	96622693
